refactor:SSR / CSR 분리-#136
